original_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
axial-positional-embedding==0.3.12
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
braindecode==1.0.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colt5-attention==0.11.1
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.6.0
cudf-polars-cu12==25.6.0
cupy-cuda12x==13.5.1
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.24.0
dill==0.3.8
docstring-inheritance==2.2.2
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.17.3
hyper-connections==0.2.1
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lark==1.2.2
lazy-loader==0.4
libcudf-cu12==25.6.0
libkvikio-cu12==25.6.0
librmm-cu12==25.6.0
linear-attention-transformer==0.19.1
linformer==0.2.3
llvmlite==0.44.0
local-attention==1.11.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mne==1.10.0
mne-bids==0.16.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.4
notebook-shim==0.2.4
numba==0.61.2
numba-cuda==0.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.12
open-clip-torch==3.0.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
polars==1.28.1
pooch==1.8.2
product-key-memory==0.2.11
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pylibcudf-cu12==25.6.0
pynvjitlink-cu12==0.7.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.1
rapids-logger==0.1.1
referencing==0.36.2
reformer-pytorch==1.4.4
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rmm-cu12==25.6.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.11.2
send2trash==1.8.3
sentry-sdk==2.34.1
setuptools==80.9.0
six==1.17.0
skorch==1.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.14.1
torch==2.7.1
torchaudio==2.7.1
torchinfo==1.8.0
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.34.0
triton==3.3.1
types-python-dateutil==2.9.0.20250708
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wfdb==4.3.0
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
